requirements/base.txt
aiocache==0.11.1
aioredis==2.0.1
alembic~=1.7.7
boto3~=1.23.5
colorama~=0.4.4
Deprecated~=1.2.13
emails~=0.6
email-validator~=1.1.3
fastapi~=0.75.1
fastapi-mail~=0.3.5.0
gunicorn~=20.0.4
itsdangerous==1.1.0
more-itertools~=8.7.0
passlib[bcrypt]==1.7.4
psycopg2-binary~=2.9.3
pydantic[email]~=1.7.3
pyjwt==2.4.0  # Todo: Not sure if we need this one
python-dotenv~=0.20.0
python-jose[cryptography]==3.3.0
python-rapidjson==1.6
pytz>=2022.1
SQLAlchemy~=1.4.34
SQLAlchemy-Searchable~=1.4.1
structlog~=21.5.0
typer-cli~=0.0.12
uvicorn[standard]~=0.17.6
mangum>=0.11.0